Updated the langlink.php script
git-svn-id: https://subversion.assembla.com/svn/akeeba-release-system/trunk@527 5fbba189-5f39-4019-a16b-6abf1605764a

+++ langlink.php
@@ -19,6 +19,11 @@ function doTheHippyHippyShake($root, $target)
 			$from = $target.'/'.$area.'/'.$module.'/en-GB';
 			$to = $moduleDir.'/language/en-GB';
 			
+			if(!is_dir($from)) {
+				// Some things may be untranslated
+				continue;
+			}
+			
 			if(is_link($to)) {
 				if(!@unlink($to)) {
 					continue;
